// frontend/src/libs/api/generated/orval/model/getBandRecruitmentsParams.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Muslog API
 * OpenAPI spec version: 1.0.0
 */

export type GetBandRecruitmentsParams = {
/**
 * Search keyword for title or description.
 */
keyword?: string;
genre?: string;
location?: string;
status?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 */
perPage?: number;
};